edges and offsets switched to uint64_t

+++ include/CSRGraph.h
@@ -10,18 +10,18 @@
 
 // 'Compressed sparse row' implementation of Graph
 class CSRGraph {
-    int *edges, *offsets;
-    int* weights;
-    int *delta_e;
-    int *delta_w;
+    uint64_t *edges, *offsets;
+    uint64_t* weights;
+    uint64_t *delta_e;
+    uint64_t *delta_w;
     int cur_idx, v, e;
 
 public:
     CSRGraph(int v, int e) : v(v), e(e) {
-        edges = new int[e];
-        offsets = new int[v + 2];
+        edges = new uint64_t[e];
+        offsets = new uint64_t[v + 2];
         cur_idx = 0;
-        weights = new int[e];
+        weights = new uint64_t[e];
     }
 
     ~CSRGraph(){
@@ -37,11 +37,11 @@ public:
 
     void finished();
 
-    int* begin(int cur_vertex);
+    uint64_t* begin(int cur_vertex);
 
-    int* end(int cur_vertex);
+    uint64_t* end(int cur_vertex);
 
-    int* begin_weights(int cur_vertex);
+    uint64_t* begin_weights(int cur_vertex);
 };
